Changes to support firebase security
Do to limited data available during firebase auth, we have to use user.id as the key for the user object. The drawback to this is the id happens to be the users e-mail. This change sets up all the ownership references to use an id that can be public.